+++ src/pystra/reliability/importance_sampling.py
@@ -51,6 +51,7 @@ class ImportanceSampling(CrudeMonteCarlo):
             form = FORM(
                 self.model,
                 self.limit_state,
+                on_failure="return",
                 options=FORMOptions(
                     block_size=self.options.block_size,
                     transform=self.options.transform,